[Refactor] Users enum들을 enums 폴더로 묶음
